[ES|QL] Handle properly recommended queries with one pipe
