+++ src/platform/packages/shared/kbn-esql-ast/src/commands_registry/options/recommended_queries/index.ts
@@ -10,35 +10,12 @@ import { i18n } from '@kbn/i18n';
 import type { RecommendedQuery, RecommendedField } from '@kbn/esql-types';
 import type { GetColumnsByTypeFn, ISuggestionItem } from '../../types';
 import { METADATA_FIELDS } from '../metadata';
+import { prettifyQueryTemplate, prettifyQuery } from './utils';
 
 export interface EditorExtensions {
   recommendedQueries: RecommendedQuery[];
   recommendedFields: RecommendedField[];
 }
-
-const prettifyQuery = (src: string): string => {
-  // Split by pipes and format each command on its own line with indentation
-  const parts = src.split('|').map((part) => part.trim());
-
-  return parts
-    .map((part, index) => {
-      if (index === 0) {
-        // First part (FROM command) - no leading pipe or indentation
-        return part;
-      } else {
-        // All subsequent commands start with pipe and have 2-space indentation
-        return `  | ${part}`;
-      }
-    })
-    .join('\n');
-};
-
-const prettifyQueryTemplate = (query: string) => {
-  const formattedQuery = prettifyQuery(query);
-  // remove the FROM command if it exists
-  const queryParts = formattedQuery.split('|');
-  return `\n|${queryParts.slice(1).join('|')}`;
-};
 
 // Order starts with the simple ones and goes to more complex ones
